requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning and AI dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
numpy>=1.24.3
torch>=2.0.0
transformers>=4.30.0

# Language processing dependencies
langdetect==1.0.9
deep-translator==1.11.4

# Existing PDF processing dependencies
pymupdf
pillow
pytesseract
pdfplumber
regex
nltk

# Utility dependencies
tqdm
python-multipart==0.0.6
requests==2.31.0

# Development dependencies
pytest==7.4.3
httpx==0.25.2

# Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
